Searching pubmed
In command line, typing "python search.py <query>" will then return a
list of the 20 most recent articles for that query from pubmed. Going to
look into how to make it more than 20 articles ( not sure if this is
necessary though ) and to sort them by something other than recent,
probably relevance. With any luck there'll be a way to choose how you
want it sorted while sending the query to pubmed. Will look into it.